Feature criptography

+++ src/packages/server/storage_manager.py
@@ -18,6 +18,26 @@ class StorageManager():
         os.makedirs(f"{self.__path}/data", exist_ok=True)
         os.makedirs(f"{self.__path}/data/images", exist_ok=True)
     
+    def remove_images(self, username: str) -> None:
+        """Removes all images from a user"""
+        users = os.listdir(f"{self.__path}/data/images")
+
+        if username not in users:
+            raise ValueError("User not found")
+
+        years = os.listdir(f"{self.__path}/data/images/{username}")
+        for year in years:
+            months = os.listdir(f"{self.__path}/data/images/{username}/{year}")
+            for month in months:
+                days = os.listdir(f"{self.__path}/data/images/{username}/{year}/{month}")
+                for day in days:
+                    images = os.listdir(f"{self.__path}/data/images/{username}/{year}/{month}/{day}")
+                    for image in images:
+                        os.remove(f"{self.__path}/data/images/{username}/{year}/{month}/{day}/{image}")
+                    os.rmdir(f"{self.__path}/data/images/{username}/{year}/{month}/{day}")
+                os.rmdir(f"{self.__path}/data/images/{username}/{year}/{month}")
+            os.rmdir(f"{self.__path}/data/images/{username}/{year}")
+        os.rmdir(f"{self.__path}/data/images/{username}")
 
     def get_users(self) -> list:
         """Returns the list of users
@@ -130,7 +150,7 @@ class StorageManager():
 
         images = []
         for choice in choices:
-            img = ImgPackage(author=choice["user"], date=choice["date"], time=choice["time"], path=choice["path"])
+            img = ImgPackage(author=choice["user"], date=choice["date"], time=choice["time"], path=choice["path"], image=Image.open(choice["path"]))
             images.append(img)
         return images
         
@@ -170,7 +190,7 @@ class StorageManager():
 
         images = []
         for choice in choices:
-            img = ImgPackage(author=username, date=choice["date"], time=choice["time"], path=choice["path"])
+            img = ImgPackage(author=username, date=choice["date"], time=choice["time"], path=choice["path"], image=Image.open(choice["path"]))
             images.append(img)
         return images
 
@@ -249,7 +269,7 @@ class StorageManager():
 
         images = []
         for choice in choices:
-            img =ImgPackage(author=username, date = choice["date"], time=choice["time"], path=choice["path"] )
+            img =ImgPackage(author=username, date = choice["date"], time=choice["time"], path=choice["path"], image=Image.open(choice["path"]))
             images.append(img)
         return images
 
@@ -285,6 +305,8 @@ class StorageManager():
         path = os.path.dirname(path)
         if os.path.exists(path):
             while os.path.exists(path) and not os.listdir(path):
+                if path == f"{self.__path}/data/images":
+                    break
                 os.rmdir(path)
                 print(f"removed dir: {path}")
                 path = os.path.dirname(path)
